requirements_mqtt.txt
paho-mqtt==1.6.1
pymongo==4.6.1
python-dotenv==1.0.1








